Switch to 32 bit hashes on all platforms + upgrade to 64 bit on windows builds

+++ GloryEngine/Editor/Extensions/EntityScenesEditorExtension/AddComponentAction.cpp
@@ -2,7 +2,7 @@
 
 namespace Glory::Editor
 {
-	AddComponentAction::AddComponentAction(size_t typeHash, UUID uuid, size_t componentIndex) : m_ComponentTypeHash(typeHash), m_ComponentUUID(uuid), m_ComponentIndex(componentIndex)
+	AddComponentAction::AddComponentAction(uint32_t typeHash, UUID uuid, size_t componentIndex) : m_ComponentTypeHash(typeHash), m_ComponentUUID(uuid), m_ComponentIndex(componentIndex)
 	{
 	}
