+++ GloryEngine/Editor/GloryEditor/PropertyDrawer.h
@@ -22,13 +22,13 @@ namespace Glory::Editor
 	class PropertyDrawer
 	{
 	public:
-		GLORY_EDITOR_API PropertyDrawer(size_t typeHash);
+		GLORY_EDITOR_API PropertyDrawer(uint32_t typeHash);
 		virtual GLORY_EDITOR_API ~PropertyDrawer();
-		virtual GLORY_EDITOR_API bool Draw(const SerializedProperty* serializedProperty, const std::string& label, void* data, size_t typeHash, uint32_t flags) const;
+		virtual GLORY_EDITOR_API bool Draw(const SerializedProperty* serializedProperty, const std::string& label, void* data, uint32_t typeHash, uint32_t flags) const;
 		virtual GLORY_EDITOR_API bool Draw(const SerializedProperty* serializedProperty, const std::string& label, std::any& data, uint32_t flags) const;
-		virtual GLORY_EDITOR_API bool Draw(const std::string& label, std::vector<char>& buffer, size_t typeHash, size_t offset, size_t size, uint32_t flags) const;
-		virtual GLORY_EDITOR_API bool Draw(const std::string& label, void* data, size_t typeHash, uint32_t flags) const;
-		virtual GLORY_EDITOR_API bool Draw(const std::string& label, YAML::Node& node, size_t typeHash, uint32_t flags) const;
+		virtual GLORY_EDITOR_API bool Draw(const std::string& label, std::vector<char>& buffer, uint32_t typeHash, size_t offset, size_t size, uint32_t flags) const;
+		virtual GLORY_EDITOR_API bool Draw(const std::string& label, void* data, uint32_t typeHash, uint32_t flags) const;
+		virtual GLORY_EDITOR_API bool Draw(const std::string& label, YAML::Node& node, uint32_t typeHash, uint32_t flags) const;
 		GLORY_EDITOR_API bool Draw(const SerializedProperty* serializedProperty) const;
 
 		template<class T>
@@ -39,16 +39,16 @@ namespace Glory::Editor
 
 		static GLORY_EDITOR_API void RegisterPropertyDrawer(PropertyDrawer* pDrawer);
 
-		static GLORY_EDITOR_API bool DrawProperty(const SerializedProperty* serializedProperty, const std::string& label, void* data, size_t typeHash, size_t elementTypeHash, uint32_t flags);
+		static GLORY_EDITOR_API bool DrawProperty(const SerializedProperty* serializedProperty, const std::string& label, void* data, uint32_t typeHash, uint32_t elementTypeHash, uint32_t flags);
 		static GLORY_EDITOR_API bool DrawProperty(const std::string& label, std::any& data, uint32_t flags);
 		static GLORY_EDITOR_API bool DrawProperty(const SerializedProperty* serializedProperty);
-		static GLORY_EDITOR_API bool DrawProperty(const std::string& label, std::vector<char>& buffer, size_t typeHash, size_t offset, size_t size, uint32_t flags);
-		static GLORY_EDITOR_API bool DrawProperty(const std::string& label, void* data, size_t typeHash, uint32_t flags);
+		static GLORY_EDITOR_API bool DrawProperty(const std::string& label, std::vector<char>& buffer, uint32_t typeHash, size_t offset, size_t size, uint32_t flags);
+		static GLORY_EDITOR_API bool DrawProperty(const std::string& label, void* data, uint32_t typeHash, uint32_t flags);
 		static GLORY_EDITOR_API bool DrawProperty(const GloryReflect::FieldData* pFieldData, void* data, uint32_t flags);
 		static GLORY_EDITOR_API bool DrawProperty(const std::string& label, const GloryReflect::TypeData* pTypeData, void* data, uint32_t flags);
 		static GLORY_EDITOR_API bool DrawProperty(const ScriptProperty& scriptProperty, YAML::Node& node, uint32_t flags);
 
-		static GLORY_EDITOR_API PropertyDrawer* GetPropertyDrawer(size_t typeHash);
+		static GLORY_EDITOR_API PropertyDrawer* GetPropertyDrawer(uint32_t typeHash);
 
 		static GLORY_EDITOR_API const std::filesystem::path& GetCurrentPropertyPath();
 		static GLORY_EDITOR_API const GloryReflect::TypeData* GetRootTypeData();
@@ -64,7 +64,7 @@ namespace Glory::Editor
 
 	private:
 		friend class MainEditor;
-		size_t m_TypeHash;
+		uint32_t m_TypeHash;
 
 		static std::vector<PropertyDrawer*> m_PropertyDrawers;
 		static const GloryReflect::TypeData* m_pRootTypeData;
@@ -79,7 +79,7 @@ namespace Glory::Editor
 		SimplePropertyDrawerTemplate() : PropertyDrawer(ResourceType::GetHash<PropertyType>()) {}
 
 	protected:
-		virtual bool Draw(const std::string& label, void* data, size_t typeHash, uint32_t flags) const override
+		virtual bool Draw(const std::string& label, void* data, uint32_t typeHash, uint32_t flags) const override
 		{
 			PropertyType oldValue = *(PropertyType*)data;
 			if (OnGUI(label, (PropertyType*)data, flags))
@@ -94,7 +94,7 @@ namespace Glory::Editor
 			return false;
 		}
 
-		virtual bool Draw(const std::string& label, std::vector<char>& buffer, size_t typeHash, size_t offset, size_t size, uint32_t flags) const override
+		virtual bool Draw(const std::string& label, std::vector<char>& buffer, uint32_t typeHash, size_t offset, size_t size, uint32_t flags) const override
 		{
 			PropertyType value;
 			memcpy((void*)&value, (void*)&buffer[offset], size);
@@ -118,7 +118,7 @@ namespace Glory::Editor
 		PropertyDrawerTemplate() : PropertyDrawer(ResourceType::GetHash<PropertyType>()) {}
 
 	protected:
-		virtual bool Draw(const SerializedProperty* serializedProperty, const std::string& label, void* data, size_t typeHash, uint32_t flags) const override
+		virtual bool Draw(const SerializedProperty* serializedProperty, const std::string& label, void* data, uint32_t typeHash, uint32_t flags) const override
 		{
 			PropertyType oldValue = *(PropertyType*)data;
 			if (OnGUI(label, (PropertyType*)data, flags))
@@ -130,7 +130,7 @@ namespace Glory::Editor
 			return false;
 		}
 
-		virtual bool Draw(const std::string& label, void* data, size_t typeHash, uint32_t flags) const override
+		virtual bool Draw(const std::string& label, void* data, uint32_t typeHash, uint32_t flags) const override
 		{
 			PropertyType oldValue = *(PropertyType*)data;
 			if (OnGUI(label, (PropertyType*)data, flags))
@@ -159,7 +159,7 @@ namespace Glory::Editor
 			return result;
 		}
 
-		virtual bool Draw(const std::string& label, std::vector<char>& buffer, size_t typeHash, size_t offset, size_t size, uint32_t flags) const override
+		virtual bool Draw(const std::string& label, std::vector<char>& buffer, uint32_t typeHash, size_t offset, size_t size, uint32_t flags) const override
 		{
 			PropertyType value;
 			memcpy((void*)&value, (void*)&buffer[offset], size);
@@ -173,7 +173,7 @@ namespace Glory::Editor
 			return true;
 		}
 
-		virtual bool Draw(const std::string& label, YAML::Node& node, size_t typeHash, uint32_t flags) const override
+		virtual bool Draw(const std::string& label, YAML::Node& node, uint32_t typeHash, uint32_t flags) const override
 		{
 			PropertyType oldValue = PropertyType();
 			if (node.IsDefined()) oldValue = node.as<PropertyType>();
